backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0
sqlalchemy[asyncio]
anthropic>=0.18.0
google-generativeai>=0.3.0